Moved image encoding functions to upper `Files` directory
Will be generalized for any file type, not just images
